requirements.txt
tensorflow
opencv-python
pillow
tk
